Square root

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -7,7 +7,7 @@ int squareroot(int, int);
  */
 int _sqrt_recursion(int n)
 {
-	unsigned int b;
+	long int b;
 
 	if (n == 1)
 		return (1);
